🦄 refactor: 新增model层
